combine legacy script into route
